services/auth/requirements.txt
flask
flask-jwt-extended
flask-cors
python-dotenv
pyyaml
jsonschema
flask-swagger-ui
